Fix broken binary stls when long is 64bit
This happens on macOS both for x86_64 and arm64

+++ source/writer.cpp
@@ -152,7 +152,7 @@ void StlWriter::writeStlBinary()
 		wfile << c;
 	}
 	TriMeshPtr mesh = triMesh();
-	long trisize = mesh->sizeTriangles();
+	uint32_t trisize = mesh->sizeTriangles();
 	wfile.write((char *)(&trisize), sizeof(trisize));
 
 	TriVIterator iter;
